+++ ft-sdk/src/main/java/com/ft/sdk/ExtraLogCatSetting.java
@@ -3,7 +3,7 @@ package com.ft.sdk;
 /**
  * 额外 log cat日志配置
  */
-public class ExtraLogCatWithError {
+public class ExtraLogCatSetting {
 
     private static final int MINI_LINES = 0;
     private static final int MAX_LINES = 500;
@@ -12,17 +12,23 @@ public class ExtraLogCatWithError {
     private int logcatSystemLines = 50;
     private int logcatEventsLines = 50;
 
+    /**
+     * logcatMainLines 默认 200
+     * logcatSystemLines 默认 50
+     * logcatEventsLines 默认 50
+     * 自定义，请使用 {@link #ExtraLogCatSetting(int, int, int)}
+     */
 
-    public ExtraLogCatWithError() {
+    public ExtraLogCatSetting() {
 
     }
 
     /**
-     * @param logcatMainLines   这是主要的日志缓冲区，包含大部分应用程序的日志输出，[0,500]，default 200
-     * @param logcatSystemLines 系统日志缓冲区，包含系统级别的日志信息，[0,500]，default 50
-     * @param logcatEventsLines 事件日志缓冲区，主要记录特定的事件信息，[0,500]，default 50
+     * @param logcatMainLines   这是主要的日志缓冲区，包含大部分应用程序的日志输出，[0,500]，默认 200
+     * @param logcatSystemLines 系统日志缓冲区，包含系统级别的日志信息，[0,500]，默认 50
+     * @param logcatEventsLines 事件日志缓冲区，主要记录特定的事件信息，[0,500]，默认 50
      */
-    public ExtraLogCatWithError(int logcatMainLines, int logcatSystemLines, int logcatEventsLines) {
+    public ExtraLogCatSetting(int logcatMainLines, int logcatSystemLines, int logcatEventsLines) {
         this.logcatMainLines = Math.min(Math.max(logcatMainLines, MINI_LINES), MAX_LINES);
         this.logcatSystemLines = Math.min(Math.max(logcatSystemLines, MINI_LINES), MAX_LINES);
         this.logcatEventsLines = Math.min(Math.max(logcatEventsLines, MINI_LINES), MAX_LINES);