+++ ft-sdk/src/main/java/com/ft/sdk/FTRUMConfig.java
@@ -83,9 +83,19 @@ public class FTRUMConfig {
     private String serviceName = Constants.DEFAULT_SERVICE_NAME;
 
     /**
-     * Crash 时 logcat 附属信息
+     * Java Crash 时 logcat 附属信息
      */
-    private ExtraLogCatWithError extraLogCatWithError;
+    private ExtraLogCatSetting extraLogCatWithJavaCrash;
+
+    /**
+     * Native Crash 时 logcat 附属信息
+     */
+    private ExtraLogCatSetting extraLogCatWithNativeCrash;
+
+    /**
+     * ANR  时 logcat 附属信息
+     */
+    private ExtraLogCatSetting extraLogCatWithANR;
 
     /**
      * 获取采样率
@@ -157,14 +167,17 @@ public class FTRUMConfig {
     /**
      * 设置是否监测 App 崩溃
      *
-     * @param enableTrackAppCrash 是否开启 Crash 最总
-     * @param  extraLogCatWithError  Java Crash 附加配置 logcat
+     * @param enableTrackAppCrash        是否开启 Crash 最总
+     * @param extraLogCatWithJavaCrash   Java Crash 附加配置 logcat
+     * @param extraLogCatWithNativeCrash Native Crash 附加配置 logcat
      * @return
      */
     public FTRUMConfig setEnableTrackAppCrash(boolean enableTrackAppCrash,
-                                              ExtraLogCatWithError extraLogCatWithError) {
+                                              ExtraLogCatSetting extraLogCatWithJavaCrash,
+                                              ExtraLogCatSetting extraLogCatWithNativeCrash) {
         this.enableTrackAppCrash = enableTrackAppCrash;
-        this.extraLogCatWithError = extraLogCatWithError;
+        this.extraLogCatWithJavaCrash = extraLogCatWithJavaCrash;
+        this.extraLogCatWithNativeCrash = extraLogCatWithNativeCrash;
         return this;
     }
 
@@ -187,8 +200,28 @@ public class FTRUMConfig {
         return enableTrackAppANR;
     }
 
+    /**
+     * 是否开启 ANR 检测，默认为 false
+     *
+     * @param enableTrackAppANR
+     * @return
+     */
+
     public FTRUMConfig setEnableTrackAppANR(boolean enableTrackAppANR) {
         this.enableTrackAppANR = enableTrackAppANR;
+        return this;
+    }
+
+    /**
+     * 是否开启 ANR 检测，默认为 false
+     *
+     * @param enableTrackAppANR
+     * @param extraLogCatWithANR ANR Crash 附加配置 logcat
+     * @return
+     */
+    public FTRUMConfig setEnableTrackAppANR(boolean enableTrackAppANR, ExtraLogCatSetting extraLogCatWithANR) {
+        this.enableTrackAppANR = enableTrackAppANR;
+        this.extraLogCatWithANR = extraLogCatWithANR;
         return this;
     }
 
@@ -342,10 +375,17 @@ public class FTRUMConfig {
         return this.enableResourceHostIP;
     }
 
-    public ExtraLogCatWithError getExtraLogCatWithError() {
-        return extraLogCatWithError;
+    public ExtraLogCatSetting getExtraLogCatWithJavaCrash() {
+        return extraLogCatWithJavaCrash;
     }
 
+    public ExtraLogCatSetting getExtraLogCatWithNativeCrash() {
+        return extraLogCatWithNativeCrash;
+    }
+
+    public ExtraLogCatSetting getExtraLogCatWithANR() {
+        return extraLogCatWithANR;
+    }
 
     //    /**
 //     * 设置 BackendSample 后端采样，当为 true 时，rum sampleRate 设置不再起效